app/requirements.txt
asgiref==3.5.0
Django==4.0.10
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sqlparse==0.4.2
uritemplate==4.1.1
